ipsec: libreswan: Fix regexp for connections waiting on child SA.
These should be considered active, because pluto is waiting for the
other side to react.  We should not remove them or try to repair.
Such connections have an extra text between the SA number and the
name of the connection.

Ideally, we would like not to parse the output of ipsec status, since
it's very error prone, but there is, unfortunately, no other interface.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>