json: Use JSON maps instead of string hashes for objects.
Use of shash requires copying of the keys every time a new JSON object
is constructed.  We're spending a lot of CPU and memory resources on
copying column names into objects for every database row, we copy
all the sets between database datum and JSON objects, and we copy all
the keys in maps.

Let's use jsmap instead of an shash as a base data structure for JSON
objects.  That will allow sharing keys between the ovsdb_datum and
JSON objects using reference counting, because keys will be now JSON
structures (JSON_STRING) instead of plain byte arrays.

This commit, however, is only a ground work replacing the base data
structure and updating all the users.  No functional changes are
done here.  If anything, it can make performance a little worse until
the optimizations are done in the next commits.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>